moving attribute binding analyzers to Functions namespace
