+++ src/Analyzers/Functions/AttributeBinding/DurableClientBindingAnalyzer.cs
@@ -5,7 +5,7 @@ using System.Collections.Immutable;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
 
-namespace Microsoft.DurableTask.Analyzers.AttributeBinding;
+namespace Microsoft.DurableTask.Analyzers.Functions.AttributeBinding;
 
 /// <summary>
 /// Analyzer that matches 'DurableClientAttribute' with 'DurableTaskClient' parameters.
